+++ include/core/io/TextReader.h
@@ -17,20 +17,20 @@ namespace core { namespace io {
             /*
              * Reads a line of characters from the text reader and returns the data as a string.
              */
-            virtual int32_t readLine(char* buffer, int limit) {
+            virtual size_t readLine(char* buffer, int limit) {
                 return readToEnd(buffer, '\n', limit);
             };
 
             /*
              * Reads all characters from the current position to the end of the text reader and returns them as one string.
              */
-            virtual int32_t readToEnd(char* buffer, int limit) {
+            virtual size_t readToEnd(char* buffer, int limit) {
                 return readToEnd(buffer, '\0', limit);
             };
 
-            virtual int32_t readToEnd(char* buffer, char endMark, int32_t limit) {
+            virtual size_t readToEnd(char* buffer, char endMark, size_t limit) {
                 int32_t r = -1;
-                int32_t c = 0;
+                size_t c = 0;
                 while (c < limit && (r = read()) != -1 && r != (int32_t)endMark) {
                     buffer[c] = (char) r;
                     c++;
